+++ CHANGELOG.rst
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`_
 and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_
 
+[unreleased]
+------------
+Changed
+^^^^^^^
+- change loop argument "params" to "parameters"
+
 [0.1.1] - 2022-04-06
 --------------------
 Added